telegram_bot/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
diskcache==5.6.1
executing==1.2.0
frozenlist==1.3.3
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
llama-cpp-python==0.1.68
magic-filter==1.0.9
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.25.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.6
yarl==1.9.2
